% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidywizard_app.R
\docType{data}
\name{intro}
\alias{intro}
\title{The main tidywizard app}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
intro
}
\value{
Defines the tidywizard shiny application
}
\description{
The main tidywizard app
}
\keyword{datasets}
