% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fcts.R
\name{logical_condition}
\alias{logical_condition}
\title{Helper functions for the tidywizard app}
\usage{
logical_condition(input_condition)
}
\value{
useful functions for tidywizard
}
\description{
Helper functions for the tidywizard app
}
